✨ CORE: Refactor Helios
- Extracted `Helios` class and types to `packages/core/src/Helios.ts`
- Updated `packages/core/src/index.ts` to export from `Helios.ts`
- Updated internal imports in `ai.ts` and `render-session.ts`
- Bumped `packages/core` version to 3.6.0
- Verified build and tests pass
